Add the Terminate method to the connection interface.
To avoid race condition and deadlock during destroing object of TcpServerOriginatedSocketConnection the Terminate method was added to the ThreadedSocketConnection.
Race condition can happen between the call FindEstablishedConnection from TcpTransportAdapter::Store and Terminate from TransportAdapterImpl.
In case when TcpServerOriginatedSocketConnection destroyed after FindEstablishedConnection called  we will get deadlock , becuase in destructor of TcpServerOriginatedSocketConnection we call StopAndJoinThread(), but now we are in the same thread.